% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fts_horizon.R
\name{get_all_fts}
\alias{get_all_fts}
\title{Get all available FTS data combined}
\usage{
get_all_fts(country = NULL)
}
\arguments{
\item{country}{Optional beneficiary country name. If NULL, includes all countries.}
}
\value{
A combined tibble of FTS data for all years
}
\description{
Get all available FTS data combined
}
