% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fts_horizon.R
\name{get_fts_data}
\alias{get_fts_data}
\title{Download and parse FTS data for a given year}
\usage{
get_fts_data(year, country = NULL)
}
\arguments{
\item{year}{An integer year between 2007 and 2024}

\item{country}{Optional beneficiary country name (e.g., "Finland"). If NULL, includes all countries.}
}
\value{
A tibble of filtered and cleaned FTS data
}
\description{
Download and parse FTS data for a given year
}
