% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fts_horizon.R
\name{get_horizon_2020}
\alias{get_horizon_2020}
\title{Load Horizon 2020 data}
\usage{
get_horizon_2020(country = NULL)
}
\arguments{
\item{country}{Optional ISO2 country code (e.g., "FI"). If NULL, includes all countries.}
}
\value{
A tibble of Horizon 2020 projects with organizations from the selected country or all countries
}
\description{
Load Horizon 2020 data
}
