% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fts_horizon.R
\name{get_horizon_europe}
\alias{get_horizon_europe}
\title{Load Horizon Europe data}
\usage{
get_horizon_europe(country = NULL)
}
\arguments{
\item{country}{Optional ISO2 country code (e.g., "FI"). If NULL, includes all countries.}
}
\value{
A tibble of Horizon Europe projects with organizations from the selected country or all countries
}
\description{
Load Horizon Europe data
}
