% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interreg.R
\name{get_interreg_data}
\alias{get_interreg_data}
\title{Retrieve Interreg project data from keep.eu API}
\usage{
get_interreg_data(country = NULL)
}
\arguments{
\item{country}{A character string or vector indicating the country name(s) or code(s) to filter results by.
If `NULL`, data for all countries is returned.}
}
\value{
A list with two elements:
\describe{
  \item{partners}{A tibble containing partner data, optionally filtered by country.}
  \item{projects}{A tibble containing project data linked to the filtered partners.}
}
}
\description{
This function sends a POST request to the [keep.eu](https://keep.eu) API and downloads an Excel file containing data on Interreg projects and their partners.
The results can be filtered by country if specified.
}
\examples{
\dontrun{
  # Retrieve all projects and partners
  data <- get_interreg_data()

  # Retrieve only projects related to Finland
  data <- get_interreg_data(country = "Finland")
}

}
